\name{FinePop-package}
\alias{FinePop-package}
\docType{package}
\title{\packageTitle{FinePop}}
\description{\packageDescription{FinePop}}
\author{
\packageAuthor{FinePop}\cr
Maintainer: \packageMaintainer{FinePop}
}
\references{
Kitada S, Kitakado T, Kishino H (2007) Empirical Bayes inference of pairwise FST and its distribution in the genome. \emph{Genetics}, 177, 861-873.\cr\cr
Kitada S, Nakamichi R, Kishino H (2017) The empirical Bayes estimators of fine-scale population structure in high gene flow species. \emph{Mol. Ecol. Resources}, DOI: 10.1111/1755-0998.12663 \cr\cr
Rousset F (2008) Genepop'007: a complete reimplementation of the Genepop software for Windows and Linux. \emph{Mol. Ecol. Resources}, 8, 103-106.\cr\cr
}
\keyword{ package }
