\name{FinePop-package}
\alias{FinePop-package}
\docType{package}
\title{\packageTitle{FinePop}}
\description{\packageDescription{FinePop}}
\author{
\packageAuthor{FinePop}\cr
Maintainer: \packageMaintainer{FinePop}
}
\references{
Kitada S, Kitakado T, Kishino H (2007) Empirical Bayes inference of pairwise FST and its distribution in the genome. Genetics, 177, 861-873.\cr\cr
Rousset F (2008) Genepop'007: a complete reimplementation of the Genepop software for Windows and Linux. Mol. Ecol. Resources 8, 103-106.
}
\keyword{ package }
