\name{FinePop-package}
\alias{FinePop-package}
\docType{package}
\title{\packageTitle{FinePop}}
\description{\packageDescription{FinePop}}
\author{
\packageAuthor{FinePop}\cr
Maintainer: \packageMaintainer{FinePop}
}
\references{
Hedrick P (2005) A standardized genetic differentiation measure. \emph{Evolution}, 59, 1633-1638.\cr\cr
Jost L (2008) Gst and its relatives do not measure differentiation. \emph{Molecular Ecology}, 17, 4015-4026.\cr\cr
Kitada S, Kitakado T, Kishino H (2007) Empirical Bayes inference of pairwise FST and its distribution in the genome. \emph{Genetics}, 177, 861-873.\cr\cr
Nei M (1973) Analysis of Gene Diversity in Subdivided Populations. \emph{Proc. Nat. Acad. Sci.}, 70, 3321-3323.\cr\cr
Nei M, Chesser RK (1983) Estimation of fixation indices and gene diversity. \emph{Annals of Human Genetics}, 47, 253-259.\cr\cr
Rousset F (2008) Genepop'007: a complete reimplementation of the Genepop software for Windows and Linux. \emph{Mol. Ecol. Resources}, 8, 103-106.\cr\cr
Weir BS, Cockerham CC (1984) Estimating F-statistics for the analysis of population structure. \emph{Evolution}, 38, 1358-1370.
}
\keyword{ package }
